# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Crypto Arbitrage Scanner API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: crypto
    description: Crypto market data and arbitrage scanning
  - name: markets
    description: All-categories Polymarket browser
  - name: stocks
    description: Stock and index market data and recommendations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /crypto/binance:
    get:
      operationId: getBinanceData
      tags: [crypto]
      summary: Get Binance Futures data for a single asset
      description: Returns current mark price and funding rate for a given symbol
      parameters:
        - in: query
          name: symbol
          schema:
            type: string
            default: BTCUSDT
          required: false
          description: Trading symbol (e.g. BTCUSDT, ETHUSDT)
      responses:
        "200":
          description: Binance futures data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BinanceData"
        "502":
          description: Failed to fetch from Binance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /crypto/binance/multi:
    get:
      operationId: getBinanceMulti
      tags: [crypto]
      summary: Get Binance Futures data for multiple assets
      description: Returns current mark price and funding rate for BTC, ETH, SOL, and BNB
      responses:
        "200":
          description: Array of Binance futures data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BinanceData"
        "502":
          description: Failed to fetch from Binance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /crypto/polymarket:
    get:
      operationId: getPolymarketMarkets
      tags: [crypto]
      summary: Get active Polymarket crypto markets
      description: Returns active crypto prediction markets from Polymarket
      parameters:
        - in: query
          name: asset
          schema:
            type: string
            enum: [BTC, ETH, SOL, BNB, ALL]
            default: ALL
          required: false
          description: Filter by crypto asset
        - in: query
          name: search
          schema:
            type: string
          required: false
          description: Free-text search within question text
      responses:
        "200":
          description: List of active crypto markets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PolymarketMarket"
        "502":
          description: Failed to fetch from Polymarket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /crypto/scan:
    get:
      operationId: getScanResults
      tags: [crypto]
      summary: Run full arbitrage/sentiment scan
      description: Fetches Binance and Polymarket data, cross-references them, and returns arbitrage signals
      parameters:
        - in: query
          name: asset
          schema:
            type: string
            enum: [BTC, ETH, SOL, BNB, ALL]
            default: ALL
          required: false
          description: Asset(s) to scan
        - in: query
          name: search
          schema:
            type: string
          required: false
          description: Free-text filter on market questions
      responses:
        "200":
          description: Scan results with signals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResult"
        "502":
          description: Failed to fetch external data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /markets/all:
    get:
      operationId: getAllMarkets
      tags: [markets]
      summary: Browse all Polymarket markets (all categories)
      description: Returns all active Polymarket prediction markets across all categories — crypto, politics, sports, economy, tech, and more
      parameters:
        - in: query
          name: category
          schema:
            type: string
            enum: [ALL, CRYPTO, POLITICS, SPORTS, ECONOMY, TECH, OTHER]
            default: ALL
          required: false
          description: Filter by category
        - in: query
          name: search
          schema:
            type: string
          required: false
          description: Free-text search within question text
      responses:
        "200":
          description: List of all active markets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PolymarketMarket"
        "502":
          description: Failed to fetch from Polymarket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stocks:
    get:
      operationId: getStocks
      tags: [stocks]
      summary: Get live stock and index quotes
      description: Returns live quotes for a curated universe of ~100 stocks (tech, energy, resources, large caps) plus common indices/ETFs
      responses:
        "200":
          description: List of stock quotes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockQuote"
        "502":
          description: Failed to fetch stock data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stocks/recommendations:
    get:
      operationId: getStockRecommendations
      tags: [stocks]
      summary: Get ranked stock trade recommendations
      description: Returns momentum-based BUY/SELL recommendations across the stock universe, ranked by conviction
      responses:
        "200":
          description: Ranked stock recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockRecommendation"
        "502":
          description: Failed to build recommendations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /crypto/recommendations:
    get:
      operationId: getRecommendations
      tags: [crypto]
      summary: Get top-ranked arbitrage recommendations
      description: Returns the top actionable signals ranked by severity and confidence across all assets
      responses:
        "200":
          description: Ranked recommendations list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recommendation"
        "502":
          description: Failed to fetch external data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /crypto/movers:
    get:
      operationId: getMarketMovers
      tags: [crypto]
      summary: Get market-moving intelligence
      description: Aggregates Fear & Greed index, top gainers/losers, trending coins, and latest crypto news headlines
      responses:
        "200":
          description: Aggregated market movers data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketMovers"
        "502":
          description: Failed to fetch external data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /crypto/overview:
    get:
      operationId: getMarketOverview
      tags: [crypto]
      summary: Get the live top-50 coin universe
      description: Returns the top USDT pairs by 24h quote volume (stablecoins and leveraged tokens excluded) from the Binance public mirror
      responses:
        "200":
          description: Live coin overview
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CoinTicker"
        "502":
          description: Failed to fetch external data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /crypto/scalp:
    get:
      operationId: getScalpSignals
      tags: [crypto]
      summary: Get short-timeframe scalp signals
      description: Computes RSI, EMA(9/21), ATR and swing structure on the 15m timeframe for top coins, returning ranked LONG/SHORT/NEUTRAL signals with entry, stop-loss and take-profit
      responses:
        "200":
          description: Ranked scalp signals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScalpSignal"
        "502":
          description: Failed to fetch external data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /crypto/shortterm:
    get:
      operationId: getShortTermMarkets
      tags: [crypto]
      summary: Get short-term crypto prediction markets
      description: Returns Polymarket crypto markets resolving within 48 hours, sorted soonest-first
      responses:
        "200":
          description: Near-term crypto markets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PolymarketMarket"
        "502":
          description: Failed to fetch external data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    BinanceData:
      type: object
      required:
        - symbol
        - asset
        - markPrice
        - fundingRate
        - fundingRatePercent
        - fetchedAt
      properties:
        symbol:
          type: string
        asset:
          type: string
          description: Short asset name e.g. BTC
        markPrice:
          type: number
        fundingRate:
          type: number
        fundingRatePercent:
          type: number
        fetchedAt:
          type: string

    PolymarketMarket:
      type: object
      required:
        - question
        - yesPrice
        - noPrice
        - yesProbabilityPercent
        - targetPrice
        - conditionId
        - active
        - assetTag
        - category
      properties:
        conditionId:
          type: string
        question:
          type: string
        yesPrice:
          type: number
        noPrice:
          type: number
        yesProbabilityPercent:
          type: number
        targetPrice:
          type: ["number", "null"]
        active:
          type: boolean
        endDate:
          type: ["string", "null"]
        volume:
          type: ["number", "null"]
        volume24hr:
          type: ["number", "null"]
          description: 24-hour trading volume in USDT
        assetTag:
          type: string
          description: Which asset this market relates to e.g. BTC, ETH
        category:
          type: string
          description: "Broad category: CRYPTO, POLITICS, SPORTS, ECONOMY, TECH, OTHER"
        slug:
          type: ["string", "null"]
          description: Polymarket market slug for direct URL linking
        eventSlug:
          type: ["string", "null"]
          description: Parent event slug — the correct slug for public polymarket.com/event URLs
        oneDayPriceChange:
          type: ["number", "null"]
          description: 24-hour change in YES probability (price points, e.g. 0.05 = +5pt)
        liquidity:
          type: ["number", "null"]
          description: Market liquidity in USD
        spread:
          type: ["number", "null"]
          description: Bid-ask spread (lower = tighter, more liquid)
        competitive:
          type: ["number", "null"]
          description: Polymarket competitiveness score 0-1 (higher = more two-sided action)

    ArbitrageSignal:
      type: object
      required:
        - type
        - message
        - severity
      properties:
        type:
          type: string
          enum: [overbought_sentiment, underpriced_probability, neutral]
        message:
          type: string
        severity:
          type: string
          enum: [low, medium, high]

    MarketAnalysis:
      type: object
      required:
        - market
        - distanceToTargetPercent
        - signal
        - binanceSymbol
        - markPrice
      properties:
        market:
          $ref: "#/components/schemas/PolymarketMarket"
        distanceToTargetPercent:
          type: number
        signal:
          $ref: "#/components/schemas/ArbitrageSignal"
        binanceSymbol:
          type: string
        markPrice:
          type: number

    ScanResult:
      type: object
      required:
        - binanceAssets
        - markets
        - scannedAt
        - totalMarkets
        - signalCounts
      properties:
        binanceAssets:
          type: array
          items:
            $ref: "#/components/schemas/BinanceData"
        markets:
          type: array
          items:
            $ref: "#/components/schemas/MarketAnalysis"
        scannedAt:
          type: string
        totalMarkets:
          type: integer
        signalCounts:
          $ref: "#/components/schemas/SignalCounts"

    SignalCounts:
      type: object
      required:
        - overbought
        - underpriced
        - neutral
      properties:
        overbought:
          type: integer
        underpriced:
          type: integer
        neutral:
          type: integer

    Recommendation:
      type: object
      required:
        - rank
        - action
        - rationale
        - market
        - signal
        - binanceSymbol
        - markPrice
        - distanceToTargetPercent
        - confidence
        - edge
        - potentialReturn
        - entryPrice
      properties:
        rank:
          type: integer
        action:
          type: string
          enum: [BUY_YES, BUY_NO, WATCH]
        rationale:
          type: string
        market:
          $ref: "#/components/schemas/PolymarketMarket"
        signal:
          $ref: "#/components/schemas/ArbitrageSignal"
        binanceSymbol:
          type: string
        markPrice:
          type: number
        distanceToTargetPercent:
          type: number
        confidence:
          type: string
          enum: [HIGH, MEDIUM, LOW]
        edge:
          type: number
          description: Probability mispricing in percentage points — how far off the crowd is from rational probability
        potentialReturn:
          type: number
          description: Return multiplier if position is correct (e.g. 8.5 means 8.5x your entry)
        entryPrice:
          type: number
          description: The price to pay per contract (yesPrice for BUY_YES, noPrice for BUY_NO)

    StockQuote:
      type: object
      required:
        - symbol
        - name
        - category
        - price
        - previousClose
        - change
        - changePercent
        - currency
        - tradingViewSymbol
        - fetchedAt
      properties:
        symbol:
          type: string
        name:
          type: string
        category:
          type: string
          enum: [TECH, ENERGY, RESOURCES, LARGE_CAP, INDEX]
        price:
          type: number
        previousClose:
          type: number
        change:
          type: number
        changePercent:
          type: number
        dayHigh:
          type: ["number", "null"]
        dayLow:
          type: ["number", "null"]
        monthHigh:
          type: ["number", "null"]
        monthLow:
          type: ["number", "null"]
        momentum5dPercent:
          type: ["number", "null"]
        volume:
          type: ["number", "null"]
        currency:
          type: string
        tradingViewSymbol:
          type: string
          description: Symbol formatted for TradingView links (dot notation for class shares)
        fetchedAt:
          type: string

    StockRecommendation:
      type: object
      required:
        - rank
        - symbol
        - name
        - category
        - action
        - confidence
        - rationale
        - price
        - changePercent
        - momentum5dPercent
        - rangePositionPercent
        - tradingViewSymbol
      properties:
        rank:
          type: integer
        symbol:
          type: string
        name:
          type: string
        category:
          type: string
          enum: [TECH, ENERGY, RESOURCES, LARGE_CAP, INDEX]
        action:
          type: string
          enum: [BUY, SELL, HOLD]
        confidence:
          type: string
          enum: [HIGH, MEDIUM, LOW]
        rationale:
          type: string
        price:
          type: number
        changePercent:
          type: number
        momentum5dPercent:
          type: number
        rangePositionPercent:
          type: number
          description: Where the price sits within its 1-month range (0 = monthly low, 100 = monthly high)
        tradingViewSymbol:
          type: string

    MoverCoin:
      type: object
      required:
        - symbol
        - price
        - changePercent
      properties:
        symbol:
          type: string
        price:
          type: number
        changePercent:
          type: number
        quoteVolume:
          type: ["number", "null"]

    TrendingCoin:
      type: object
      required:
        - name
        - symbol
      properties:
        name:
          type: string
        symbol:
          type: string
        marketCapRank:
          type: ["integer", "null"]
        thumb:
          type: ["string", "null"]

    NewsItem:
      type: object
      required:
        - title
        - source
        - url
        - publishedAt
      properties:
        title:
          type: string
        source:
          type: string
        url:
          type: string
        publishedAt:
          type: string

    FearGreed:
      type: object
      required:
        - value
        - classification
      properties:
        value:
          type: integer
        classification:
          type: string

    MarketMovers:
      type: object
      required:
        - fetchedAt
        - gainers
        - losers
        - trending
        - news
      properties:
        fetchedAt:
          type: string
        fearGreed:
          oneOf:
            - $ref: "#/components/schemas/FearGreed"
            - type: "null"
        gainers:
          type: array
          items:
            $ref: "#/components/schemas/MoverCoin"
        losers:
          type: array
          items:
            $ref: "#/components/schemas/MoverCoin"
        trending:
          type: array
          items:
            $ref: "#/components/schemas/TrendingCoin"
        news:
          type: array
          items:
            $ref: "#/components/schemas/NewsItem"

    CoinTicker:
      type: object
      required:
        - symbol
        - asset
        - price
        - changePercent
        - high24h
        - low24h
        - quoteVolume
        - trades
      properties:
        symbol:
          type: string
        asset:
          type: string
        price:
          type: number
        changePercent:
          type: number
        high24h:
          type: number
        low24h:
          type: number
        quoteVolume:
          type: number
        trades:
          type: number

    ScalpSignal:
      type: object
      required:
        - symbol
        - asset
        - direction
        - confidence
        - score
        - price
        - changePercent
        - rsi
        - emaFast
        - emaSlow
        - atr
        - entry
        - stopLoss
        - takeProfit
        - riskReward
        - reasons
      properties:
        symbol:
          type: string
        asset:
          type: string
        direction:
          type: string
          enum: [LONG, SHORT, NEUTRAL]
        confidence:
          type: string
          enum: [HIGH, MEDIUM, LOW]
        score:
          type: number
        price:
          type: number
        changePercent:
          type: number
        rsi:
          type: number
        emaFast:
          type: number
        emaSlow:
          type: number
        atr:
          type: number
        entry:
          type: number
        stopLoss:
          type: number
        takeProfit:
          type: number
        riskReward:
          type: number
        reasons:
          type: array
          items:
            type: string