# lib/api-zod/src/generated/types/coinTicker.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Crypto Arbitrage Scanner API
 * OpenAPI spec version: 0.1.0
 */

export interface CoinTicker {
  symbol: string;
  asset: string;
  price: number;
  changePercent: number;
  high24h: number;
  low24h: number;
  quoteVolume: number;
  trades: number;
}